FL-274: remove remaining gentoo network tweaks which cause problems in Funtoo

+++ net-misc/networkmanager/networkmanager-0.9.6.4-r4.ebuild
@@ -2,21 +2,25 @@
 
 EAPI="4"
 GNOME_ORG_MODULE="NetworkManager"
+VALA_MIN_API_VERSION="0.18"
+VALA_USE_DEPEND="vapigen"
 
-inherit autotools eutils gnome.org linux-info systemd user
+inherit autotools eutils gnome.org linux-info systemd user toolchain-funcs vala
 
 DESCRIPTION="Network configuration and management in an easy way. Desktop environment independent."
 HOMEPAGE="http://www.gnome.org/projects/NetworkManager/"
 
 LICENSE="GPL-2+"
 SLOT="0"
-IUSE="avahi bluetooth connection-sharing dhclient +dhcpcd doc gnutls +introspection kernel_linux +nss modemmanager +ppp resolvconf systemd +wext wimax"
-KEYWORDS="~amd64 ~arm ~ppc ~ppc64 ~x86"
+IUSE="avahi bluetooth connection-sharing consolekit dhclient +dhcpcd doc gnutls +introspection kernel_linux +nss modemmanager policykit ppp resolvconf systemd vala +wext wimax"
+KEYWORDS="*"
 
 REQUIRED_USE="
 	modemmanager? ( ppp )
 	^^ ( nss gnutls )
-	^^ ( dhclient dhcpcd )"
+	^^ ( dhclient dhcpcd )
+	consolekit? ( !systemd )
+	systemd? ( !consolekit )"
 
 # gobject-introspection-0.10.3 is needed due to gnome bug 642300
 # wpa_supplicant-0.7.3-r3 is needed due to bug 359271
@@ -24,10 +28,10 @@ REQUIRED_USE="
 # XXX: on bump, check that net-wireless/wimax is still using libnl:1.1 !
 # TODO: Qt support?
 COMMON_DEPEND=">=sys-apps/dbus-1.2
-	>=dev-libs/dbus-glib-0.75
+	>=dev-libs/dbus-glib-0.94
 	|| ( >=sys-fs/udev-171[gudev] >=sys-fs/udev-147[extras] )
 	>=dev-libs/glib-2.26
-	>=sys-auth/polkit-0.97
+	policykit? ( >=sys-auth/polkit-0.97 )
 	>=net-libs/libsoup-2.26:2.4
 	>=net-wireless/wpa_supplicant-0.7.3-r3[dbus]
 	bluetooth? ( >=net-wireless/bluez-4.82 )
@@ -52,14 +56,19 @@ COMMON_DEPEND=">=sys-apps/dbus-1.2
 RDEPEND="${COMMON_DEPEND}
 	modemmanager? ( >=net-misc/modemmanager-0.4 )
 	systemd? ( >=sys-apps/systemd-31 )
-	!systemd? ( sys-auth/consolekit )"
+	consolekit? ( sys-auth/consolekit )
+	net-wireless/rfkill
+	=dev-lang/python-3*"
+
+# addwifi script requires rfkill and python ^^^
 
 DEPEND="${COMMON_DEPEND}
 	virtual/pkgconfig
 	>=dev-util/intltool-0.40
 	>=sys-devel/gettext-0.17
 	>=sys-kernel/linux-headers-2.6.29
-	doc? ( >=dev-util/gtk-doc-1.8 )"
+	doc? ( >=dev-util/gtk-doc-1.8 )
+	vala? ( $(vala_depend) )"
 
 sysfs_deprecated_check() {
 	ebegin "Checking for SYSFS_DEPRECATED support"
@@ -96,40 +105,33 @@ src_prepare() {
 	epatch "${FILESDIR}/${PN}-0.9_rc3-fix-tests.patch"
 	# Build against libnl:1.1 for net-wireless/wimax-1.5.2 compatibility
 	epatch "${FILESDIR}/${PN}-0.9.4.0-force-libnl1.1-r1.patch"
-	# Update init.d script to provide net and use inactive status if not connected
-	epatch "${FILESDIR}/${PN}-0.9.2.0-init-provide-net-r1.patch"
 	# Bug #402085, https://bugzilla.gnome.org/show_bug.cgi?id=387832
 	epatch "${FILESDIR}/${PN}-0.9.2.0-pre-sleep.patch"
-	# Fix quote handling for global data (Bug #410821)
-	epatch "${FILESDIR}/${PN}-0.9.4.0-fix-quote-handling.patch"
-	# Fix uninitialized variables in libnm-glib
-	epatch "${FILESDIR}/${P}-libnm-glib-ensure_inited.patch"
-	epatch "${FILESDIR}/${P}-libnm-glib-init-gerror.patch"
-	# Fix building against linux-headers-3.4, #417055
-	epatch "${FILESDIR}/${P}-ip_ppp.h.patch"
-	# Fix ipv6 default route bug, #417529
-	epatch "${FILESDIR}/${P}-ipv6-route.patch"
 	# Bug #335147, https://bugzilla.gnome.org/show_bug.cgi?id=679428
-	epatch "${FILESDIR}/${P}-dhclient-ipv6.patch"
-	# Bug #426844, wext vs. nl80211 API issues
-	epatch "${FILESDIR}/${P}-nl80211-encryption-caps.patch"
-	epatch "${FILESDIR}/${P}-wifi-kernel-api.patch"
+	epatch "${FILESDIR}/${PN}-0.9.4.0-dhclient-ipv6.patch"
+	# https://bugzilla.gnome.org/show_bug.cgi?id=683932
+	epatch "${FILESDIR}/${PN}-0.9.6.0-daemon-signals.patch"
 
 	epatch_user
 
+	use vala && vala_src_prepare
 	eautoreconf
 	default
 }
 
 src_configure() {
+	local udevdir=/lib/udev
+	has_version sys-fs/udev && udevdir="$($(tc-getPKG_CONFIG) --variable=udevdir udev)"
+
 	ECONF="--disable-more-warnings
 		--disable-static
 		--localstatedir=/var
 		--with-distro=gentoo
 		--with-dbus-sys-dir=/etc/dbus-1/system.d
-		--with-udev-dir=/lib/udev
+		--with-udev-dir=${udevdir}
 		--with-iptables=/sbin/iptables
 		--enable-concheck
+		$(use_enable policykit polkit)
 		$(use_enable doc gtk-doc)
 		$(use_enable introspection)
 		$(use_enable ppp)
@@ -138,6 +140,7 @@ src_configure() {
 		$(use_with dhcpcd)
 		$(use_with doc docs)
 		$(use_with resolvconf)
+		$(use_enable vala)
 		$(use_with wext)
 		$(use_with wimax libnl-1)
 		$(systemd_with_unitdir)"
@@ -150,8 +153,10 @@ src_configure() {
 
 		if use systemd; then
 			ECONF="${ECONF} --with-session-tracking=systemd"
-		else
+		elif use consolekit; then
 			ECONF="${ECONF} --with-session-tracking=ck"
+		else
+			ECONF="${ECONF} --with-session-tracking=none"
 		fi
 
 	econf ${ECONF}
@@ -165,31 +170,25 @@ src_install() {
 	# Need to keep the /etc/NetworkManager/dispatched.d for dispatcher scripts
 	keepdir /etc/NetworkManager/dispatcher.d
 
-	# Provide openrc net dependency only when nm is connected
-	exeinto /etc/NetworkManager/dispatcher.d
-	newexe "${FILESDIR}/10-openrc-status-r2" 10-openrc-status
-	sed -e "s:@EPREFIX@:${EPREFIX}:g" \
-		-i "${ED}/etc/NetworkManager/dispatcher.d/10-openrc-status" || die
-
 	# Add keyfile plugin support
 	keepdir /etc/NetworkManager/system-connections
 	chmod 0600 "${ED}"/etc/NetworkManager/system-connections/.keep* # bug #383765
 	insinto /etc/NetworkManager
 	newins "${FILESDIR}/nm-system-settings.conf-ifnet" NetworkManager.conf
 
-	# Allow users in plugdev group to modify system connections
-	insinto /usr/share/polkit-1/rules.d/
-	doins "${FILESDIR}/01-org.freedesktop.NetworkManager.settings.modify.system.rules"
-	if has_version '<sys-auth/polkit-0.106'; then
-		insinto /etc/polkit-1/localauthority/10-vendor.d
-		doins "${FILESDIR}/01-org.freedesktop.NetworkManager.settings.modify.system.pkla"
+	if use policykit; then
+		# Allow users in plugdev group to modify system connections
+		insinto /usr/share/polkit-1/rules.d/
+		doins "${FILESDIR}/01-org.freedesktop.NetworkManager.settings.modify.system.rules"
+		if has_version '<sys-auth/polkit-0.106'; then
+			insinto /etc/polkit-1/localauthority/10-vendor.d
+			doins "${FILESDIR}/01-org.freedesktop.NetworkManager.settings.modify.system.pkla"
+		fi
 	fi
-
-	# Default conf.d file
-	newconfd "${FILESDIR}/conf.d.NetworkManager" NetworkManager
 
 	# Remove useless .la files
 	find "${D}" -name '*.la' -exec rm -f {} + || die "la file removal failed"
+	dosbin ${FILESDIR}/addwifi
 }
 
 pkg_postinst() {
@@ -204,22 +203,24 @@ pkg_postinst() {
 		ewarn "After doing so, you can remove ${EROOT}etc/NetworkManager/nm-system-settings.conf"
 	fi
 
-	# The polkit rules file moved to /usr/share
-	old_rules="${EROOT}etc/polkit-1/rules.d/01-org.freedesktop.NetworkManager.settings.modify.system.rules"
-	if [[ -f "${old_rules}" ]]; then
-		case "$(md5sum ${old_rules})" in
-		  574d0cfa7e911b1f7792077003060240* )
-			# Automatically delete the old rules.d file if the user did not change it
-			elog
-			elog "Removing old ${old_rules} ..."
-			rm -f "${old_rules}" || eerror "Failed, please remove ${old_rules} manually"
-			;;
-		  * )
-			elog "The ${old_rules}"
-			elog "file moved to /usr/share/polkit-1/rules.d/ in >=networkmanager-0.9.4.0-r4"
-			elog "If you edited ${old_rules}"
-			elog "without changing its behavior, you may want to remove it."
-			;;
-		esac
+	if use policykit; then
+		# The polkit rules file moved to /usr/share
+		old_rules="${EROOT}etc/polkit-1/rules.d/01-org.freedesktop.NetworkManager.settings.modify.system.rules"
+		if [[ -f "${old_rules}" ]]; then
+			case "$(md5sum ${old_rules})" in
+			574d0cfa7e911b1f7792077003060240* )
+				# Automatically delete the old rules.d file if the user did not change it
+				elog
+				elog "Removing old ${old_rules} ..."
+				rm -f "${old_rules}" || eerror "Failed, please remove ${old_rules} manually"
+				;;
+			* )
+				elog "The ${old_rules}"
+				elog "file moved to /usr/share/polkit-1/rules.d/ in >=networkmanager-0.9.4.0-r4"
+				elog "If you edited ${old_rules}"
+				elog "without changing its behavior, you may want to remove it."
+				;;
+			esac
+		fi
 	fi
 }
